% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkOrder.R
\name{WorkOrderUI}
\alias{WorkOrderUI}
\title{序列号查询生成器界面}
\usage{
WorkOrderUI(
  tabTitle = "Work Order",
  colTitles = c("Filter Area", "Operation Area", "Result Area"),
  widthRates = c(6, 6, 12),
  func_left = WorkOrderUI_left,
  func_right = WorkOrderUI_right,
  func_bottom = WorkOrderUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
序列号查询生成器界面
}
\examples{
WorkOrderUI()
}
